Creating integrantion-test directory and moving existing integration tests to this new directory
