+++ src/integration-test/java/book/BookRestControllerTest.java
@@ -1,4 +1,4 @@
-package com.thoughtworks.librarysystem.book;
+package book;
 
 
 import com.thoughtworks.librarysystem.Application;
