+++ src/integration-test/java/library/LibraryRestControllerTest.java
@@ -1,4 +1,4 @@
-package com.thoughtworks.librarysystem.library;
+package library;
 
 
 import com.thoughtworks.librarysystem.Application;
@@ -8,6 +8,9 @@ import com.thoughtworks.librarysystem.book.BookRepository;
 import com.thoughtworks.librarysystem.copy.Copy;
 import com.thoughtworks.librarysystem.copy.CopyBuilder;
 import com.thoughtworks.librarysystem.copy.CopyRepository;
+import com.thoughtworks.librarysystem.library.Library;
+import com.thoughtworks.librarysystem.library.LibraryBuilder;
+import com.thoughtworks.librarysystem.library.LibraryRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
