+++ src/integration-test/java/loan/BorrowBookRestControllerTest.java
@@ -1,7 +1,6 @@
-package com.thoughtworks.librarysystem.loan;
+package loan;
 
 import com.thoughtworks.librarysystem.Application;
-import com.thoughtworks.librarysystem.commons.ApplicationTestBase;
 import com.thoughtworks.librarysystem.book.Book;
 import com.thoughtworks.librarysystem.book.BookBuilder;
 import com.thoughtworks.librarysystem.book.BookRepository;
@@ -15,6 +14,7 @@ import com.thoughtworks.librarysystem.library.LibraryRepository;
 import com.thoughtworks.librarysystem.user.User;
 import com.thoughtworks.librarysystem.user.UserBuilder;
 import com.thoughtworks.librarysystem.user.UserRepository;
+import commons.ApplicationTestBase;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
