+++ src/integration-test/java/user/UserRestControllerTest.java
@@ -1,58 +1,62 @@
-package com.thoughtworks.librarysystem.auth;
+package user;
 
 import com.thoughtworks.librarysystem.Application;
-import com.thoughtworks.librarysystem.commons.ApplicationTestBase;
-import com.thoughtworks.librarysystem.commons.config.AuthenticationFilter;
+import com.thoughtworks.librarysystem.user.UserRepository;
+import commons.ApplicationTestBase;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.context.WebApplicationContext;
 
-import javax.annotation.Resource;
+import java.util.HashMap;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = Application.class)
 @WebAppConfiguration
 @DirtiesContext
 @Transactional
-public class AuthTest extends ApplicationTestBase {
+public class UserRestControllerTest extends ApplicationTestBase {
 
     private MockMvc mockMvc;
 
     @Autowired
-    private AuthenticationFilter securityFilter;
+    private WebApplicationContext webApplicationContext;
 
     @Autowired
-    private WebApplicationContext webApplicationContext;
+    private UserRepository userRepository;
 
     @Before
     public void setup() throws Exception {
-
-        DefaultMockMvcBuilder defaultMockMvcBuilder = webAppContextSetup(webApplicationContext);
-        defaultMockMvcBuilder.addFilters(securityFilter);
-        this.mockMvc = defaultMockMvcBuilder.build();
-
+        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
     }
 
     @Test
-    public void shouldReturnForbiddenWhenTokenNotExists() throws Exception {
-        mockMvc.perform(get("/"))
-                .andExpect(status().isForbidden());
+    public void shouldCreateNewUser() throws  Exception{
+
+        HashMap<String, Object> inputs = new HashMap<String, Object>();
+        inputs.put("name", "Tulio Cruz");
+        inputs.put("email", "tcruz@thoughtworks.com");
+
+        String userJson = loadFixture("user.json", inputs);
+
+        mockMvc.perform(post("/users")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(userJson))
+                        .andExpect(status().isCreated());
+
     }
 
+
 }
-
-
