--- src/test/java/com/thoughtworks/librarysystem/user/UserRestControllerTest.java
@@ -1,70 +1,0 @@
-package com.thoughtworks.librarysystem.user;
-
-import com.thoughtworks.librarysystem.Application;
-import com.thoughtworks.librarysystem.book.Book;
-import com.thoughtworks.librarysystem.book.BookBuilder;
-import com.thoughtworks.librarysystem.book.BookRepository;
-import com.thoughtworks.librarysystem.commons.ApplicationTestBase;
-import com.thoughtworks.librarysystem.copy.Copy;
-import com.thoughtworks.librarysystem.copy.CopyBuilder;
-import com.thoughtworks.librarysystem.copy.CopyRepository;
-import com.thoughtworks.librarysystem.copy.CopyStatus;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.context.WebApplicationContext;
-
-import java.util.HashMap;
-
-import static org.hamcrest.Matchers.is;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = Application.class)
-@WebAppConfiguration
-@DirtiesContext
-@Transactional
-public class UserRestControllerTest extends ApplicationTestBase {
-
-    private MockMvc mockMvc;
-
-    @Autowired
-    private WebApplicationContext webApplicationContext;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Before
-    public void setup() throws Exception {
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
-    }
-
-    @Test
-    public void shouldCreateNewUser() throws  Exception{
-
-        HashMap<String, Object> inputs = new HashMap<String, Object>();
-        inputs.put("name", "Tulio Cruz");
-        inputs.put("email", "tcruz@thoughtworks.com");
-
-        String userJson = loadFixture("user.json", inputs);
-
-        mockMvc.perform(post("/users")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(userJson))
-                        .andExpect(status().isCreated());
-
-    }
-
-
-}
